WIP: Better stream CRUD implementation and enhancing exception handling

+++ app/Http/Livewire/Streams.php
@@ -4,14 +4,16 @@ namespace App\Http\Livewire;
 
 use App\Models\Stream;
 use Livewire\Component;
-use Illuminate\Support\Str;
 use Livewire\WithPagination;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Streams extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     protected $paginationTheme = 'bootstrap';
 
@@ -19,19 +21,40 @@ class Streams extends Component
 
     public $name;
     public $alias;
-    public $slug;
     public $description;
 
-    public function render()
+    public $trashed = false;
+
+    /**
+     * Lifecycle method that executes only once when the component is mounting
+     * 
+     * @param string $trashed
+     */
+    public function mount(string $trashed = null)
     {
-        return view('livewire.streams', [
-            'streams' => $this->getPaginatedLevels()
-        ]);
+        $this->trashed = boolval($trashed);
     }
 
-    public function getPaginatedLevels()
+    /**
+     * Lifecycle method method that executes everytime the component state changes
+     * 
+     * @return View
+     */
+    public function render()
     {
-        return Stream::paginate(24);
+        return view('livewire.streams', ['streams' => $this->getRelevantStreams()]);
+    }
+
+    /**
+     * Get relevant paginated streas
+     */
+    public function getRelevantStreams()
+    {
+        $streamsQuery = Stream::with('students');
+
+        if($this->trashed) $streamsQuery->onlyTrashed();
+
+        return $streamsQuery->paginate(24)->withQueryString();
     }
 
     /**
@@ -51,6 +74,11 @@ class Streams extends Component
         $this->emit('show-upsert-stream-modal');
     }
 
+    /**
+     * Streams fields validation rules
+     * 
+     * @return array
+     */
     public function rules()
     {
         return [
@@ -60,39 +88,57 @@ class Streams extends Component
         ];
     }
 
-    function createStream()
+    /**
+     * Creates a new stream database entry
+     */
+    public function createStream()
     {
-        $this->validate();
+        $data = $this->validate();
         
         try {
 
-            Stream::create([
-                'name'=>$this->name,
-               'alias'=> $this->alias,
-                'description'=>$this->description ,
-                'slug'=>Str::slug($this->name)
-            ]);
+            $this->authorize('create', Stream::class);
 
-            $this->reset(['name', 'slug', 'alias', 'description']);
+            $stream = Stream::create($data);
 
-            session()->flash('status', 'A stream has been successfully created');
+            $this->reset(['name', 'alias', 'description']);
+
+            session()->flash('status', "A stream, {$stream->name}, has been successfully created");
 
             $this->emit('hide-upsert-stream-modal');
             
         } catch (\Exception $exception) {
             
-            Log::error($exception->getMessage(), [
-                'action' => __METHOD__
-            ]);
-
-            session()->flash('error', 'A db error occurred');
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry! Adding stream operation failed");
 
             $this->emit('hide-upsert-stream-modal');
 
         }
     }
 
+    /**
+     * Set appropriate error based on the type of the error and the environment
+     * 
+     * @param \Exception $exception
+     * @param string $message
+     */
+    private function setError(\Exception $exception, string $message)
+    {
 
+        if($exception instanceof AuthorizationException) $message = $exception->getMessage();
+
+        else $message = App::environment('local') ? $exception->getMessage() : $message;
+
+        session()->flash('error', $message);
+
+    }
+
+
+    /**
+     * Updates a database stream entry
+     */
     public function updateStream()
     {
         $data = $this->validate();
@@ -102,9 +148,11 @@ class Streams extends Component
             /** @var User */
             $stream = Stream::findOrFail($this->streamId);
 
+            $this->authorize('update', $stream);
+
             if($stream->update($data)){
 
-                $this->reset(['streamId', 'name', 'slug', 'alias', 'description']);
+                $this->reset(['streamId', 'name', 'alias', 'description']);
 
                 session()->flash('status', 'stream successfully updated');
 
@@ -113,13 +161,20 @@ class Streams extends Component
             
         } catch (\Exception $exception) {
             
-            Log::error($exception->getMessage(), [
-                'action' => __METHOD__
-            ]);
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry! Updating stream operation failed");
+
+            $this->emit('hide-upsert-stream-modal');
 
         }
     }
 
+    /**
+     * Shows the delete streamm confrimation modal
+     * 
+     * @param Stream $stream
+     */
     public function showDeleteStreamModal(Stream $stream)
     {
         $this->streamId = $stream->id;
@@ -130,29 +185,32 @@ class Streams extends Component
         
     }
 
-    public function deleteLevel(Stream $stream)
+    /**
+     * Trash a stream entry
+     */
+    public function deleteStream()
     {
         try {
 
+            /** @var Stream */
             $stream = Stream::findOrFail($this->streamId);
+
+            $this->authorize('delete', $stream);
 
             if($stream->delete()){
 
                 $this->reset(['streamId', 'name']);
 
-                session()->flash('status', 'The stream has been successfully deleted');
+                session()->flash('status', "The stream, {$stream->name}, has been successfully deleted");
 
                 $this->emit('hide-delete-stream-modal');
             }
 
         } catch (\Exception $exception) {
             
-            Log::error($exception->getMessage(), [
-                'stream-id' => $this->streamId,
-                'action' => __CLASS__ . '@' . __METHOD__
-            ]);
-
-            session()->flash('error', 'A fatal error has occurred');
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry! Deleting stream operation failed");
 
             $this->emit('hide-delete-stream-modal');
         }
@@ -165,11 +223,13 @@ class Streams extends Component
     {
         try {
 
-            /** @var Collection */
-            $streams = Stream::withTrashed()->get();
+            $this->authorize('bulkDelete', Stream::class);
 
+            /** @var Collection */
+            $streams = Stream::all();
+            
             $streams->each(function(Stream $stream){
-                $stream->forceDelete();
+                $stream->delete();
             });
 
             session()->flash('status', 'All system streams have been deleted');
@@ -177,14 +237,70 @@ class Streams extends Component
             $this->emit('hide-truncate-streams-modal');
             
         } catch (\Exception $exception) {
-
-            Log::error($exception->getMessage(), [
-                'action' => __METHOD__
-            ]);
             
-            session()->flash('error', 'An error occurred while deleting systems streams');
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry! Truncating streams operation failed");
 
             $this->emit('hide-truncate-streams-modal');
         }
+    }
+
+    /**
+     * A stream can restored
+     * 
+     * @param mixed $streamId
+     */
+    public function restoreStream($streamId)
+    {
+
+        try {
+            
+            /** @var Stream */
+            $stream = Stream::where('id', $streamId)->withTrashed()->firstOrFail();
+
+            $this->authorize('restore', $stream);
+
+            $stream->restore();
+
+            session()->flash('status', "The stream, {$stream->name}, has been restored");
+
+        } catch (\Exception $exception) {
+            
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry! Restoring stream operation failed");
+            
+        }
+        
+    }
+
+    /**
+     * Completely delete a stream from the database
+     * 
+     * @param mixed $steamId
+     */
+    public function destroyStream($streamId)
+    {
+
+        try {
+            
+            /** @var Stream */
+            $stream = Stream::where('id', $streamId)->withTrashed()->firstOrFail();
+
+            $this->authorize('forceDelete', $stream);
+
+            $stream->forceDelete();
+
+            session()->flash('status', "The stream, {$stream->name}, has been completely deleted from the application");
+
+        } catch (\Exception $exception) {
+            
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry! Deleting stream operation failed, check with admin if this persistes");
+            
+        }
+        
     }
 }
